Add header for ExternalService (should be more convenient to use)
